functions/new_trip/requirements.txt
google-cloud-pubsub==2.14.1
pydantic==1.10.5